Implement the `install` and `updateGame` handlers as regular handlers
Them being in the preload files just leads to confusion (for example, it was not
obvious that the `updateGame` handler in `main.ts` was never called)
